chore: サニタイザー重複の役割を明文化+flowContextStoreテストを専用ファイル化
C-1: lib/security/inputSanitizer.ts と middleware/inputSanitizer.ts(L5)は
URL検出という関心が一部重なるが、前者は複数admin機能から使われる常時ON
のベースラインチェック、後者はchat/avatar専用のセッション単位・env切替可能な
多層防御の1層という別の関心事と判断。統合せず両ファイル冒頭にコメントで
役割の違いとドリフトリスクを明記した。

C-4: #837でTTL実装時、新規ファイル禁止のためsessionKey.test.tsと
langGraphOrchestrator.phase22.test.tsに間借りしていたflowContextStoreの
テストを、専用のflowContextStore.test.tsに移動した。
