+++ src/lib/security/inputSanitizer.ts
@@ -1,5 +1,16 @@
 // src/lib/security/inputSanitizer.ts
 // 入力サニタイズ + LLM出力クリーン
+//
+// 汎用ユーティリティ。feedback/tuning等の複数のadmin機能から素の関数として
+// 呼ばれる無条件実行のベースラインチェック（feature flag無し、常時ON）。
+// L5 (../../middleware/inputSanitizer.ts) とはURL検出という関心が一部重なるが、
+// 別物として意図的に併存させている:
+//   - こちら: 会話・非会話を問わず自由入力を受けるあらゆるエンドポイントで
+//     使える同期・ステートレスな最低限のXSS/URLガード + 出力URLストリップ
+//   - L5: chat/avatar の対話ストリーム専用。セッション単位の乱用検知・
+//     エンコーディング攻撃検知・env切り替え可能な多層防御(L5-L8)の1層
+// URL検出ロジックが両ファイルにそれぞれ実装されている点はドリフトリスクなので、
+// 片方のURLパターンだけ更新して他方を放置しないよう注意すること。
 
 // URL検出パターン
 const URL_PATTERN =
